requirements.txt
PyQt5==5.9.2
pycryptodome==3.4.7
pyperclip==1.6.0